Removing models and cron
We no longer need the models as we will not be relying on enroll file contents in the datbase for the user validation process
The cron job is no longer needed as the enroll file is bring removed
The User class and flask_login/flask_ldap3_login callabacks are being moved into __init__.py
